all tests pass, preliminary benchmarking worked
