--- tests_cupy/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2025 Antoine COLLET
-
-"""Unit test package for lbfgsb."""
