--- tests_cupy/test_cauchy.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2025 Antoine COLLET
